REO progress - binary logs

+++ java_console/ui/src/main/java/com/rusefi/ui/livedocs/LiveDocPanel.java
@@ -193,7 +193,7 @@ public class LiveDocPanel {
         combined.put(context, new RefreshActions() {
             @Override
             public void refresh(BinaryProtocol bp, byte[] response) {
-                int value = (int) conditionField.getValue(new ConfigurationImage(response));
+                int value = (int) conditionField.getValue(new ConfigurationImage(response)).intValue();
                 conditionLabel.setText(request.getField() + " is " + (value == 1 ? "TRUE" : "FALSE"));
                 JPanel active;
                 JPanel passive;
